+++ examples/data-sources/clearpass_local_user/data-source.tf
@@ -1,0 +1,7 @@
+data "clearpass_local_user" "example" {
+  id = 1234
+}
+
+output "local_user_info" {
+  value = data.clearpass_local_user.example
+}