+++ examples/data-sources/clearpass_local_users/data-source.tf
@@ -1,0 +1,6 @@
+data "clearpass_local_users" "all" {
+}
+
+output "all_local_users" {
+  value = data.clearpass_local_users.all.items
+}