+++ internal/provider/provider.go
@@ -176,5 +176,7 @@ func (p *clearpassProvider) DataSources(ctx context.Context) []func() datasource
 		NewEnforcementPoliciesDataSource,
 		NewEnforcementProfileDataSource,
 		NewEnforcementProfilesDataSource,
+		NewLocalUserDataSource,
+		NewLocalUsersDataSource,
 	}
 }